+++ webcomponents/core/src/components.d.ts
@@ -9,7 +9,8 @@ import { DeckdeckgoAttributeDefinition, DeckdeckgoDeckDefinition, DeckdeckgoSlid
 export namespace Components {
     interface DeckgoDeck {
         "animation": 'slide' | 'fade' | 'none';
-        "autoSlide": boolean;
+        "autoSlide": 'true' | 'false';
+        "autoSlideInterval": number;
         "blockSlide": (block: boolean) => Promise<void>;
         "cloneBackground": boolean;
         "deleteActiveSlide": () => Promise<void>;
@@ -94,7 +95,8 @@ declare global {
 declare namespace LocalJSX {
     interface DeckgoDeck {
         "animation"?: 'slide' | 'fade' | 'none';
-        "autoSlide"?: boolean;
+        "autoSlide"?: 'true' | 'false';
+        "autoSlideInterval"?: number;
         "cloneBackground"?: boolean;
         "direction"?: 'horizontal' | 'vertical' | 'papyrus';
         "directionMobile"?: 'horizontal' | 'vertical' | 'papyrus';